requirements.txt
tensorflow==2.12.0 
pillow>=9.0.0
requests>=2.28.0